@@ -2,38 +2,27 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LoginComponent } from './component/login/login.component';
 import { RegisterComponent } from './component/register/register.component';
-import { LandingPageComponent } from './component/landing-page/landing-page.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { DashboardLayoutComponent } from './component/dashboard-candidate/dashboard-layout/dashboard-layout/dashboard-layout.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ToastrModule } from 'ngx-toastr';
-@NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    RegisterComponent,
-   
-    
-    LandingPageComponent,
-    DashboardLayoutComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    FormsModule,
-    HttpClientModule,
-    ReactiveFormsModule,
-    BrowserAnimationsModule,
-    ToastrModule.forRoot({  
-      positionClass: 'toast-bottom-right', 
-      timeOut: 3000, 
-      preventDuplicates: true,
-    })
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+import { ToastComponent } from './component/toster/toast/toast.component';
+@NgModule({ declarations: [
+        AppComponent,
+        RegisterComponent,
+        DashboardLayoutComponent,
+        ToastComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        FormsModule,
+        ReactiveFormsModule,
+        BrowserAnimationsModule,
+        ToastrModule.forRoot({
+            positionClass: 'toast-bottom-right',
+            timeOut: 3000,
+            preventDuplicates: true,
+        })], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
